Fix stacktrace sent to browser with show_errors disabled
This commit prevents stacktraces being sent to the browser when
show_errors is disabled, to prevent potential security vulnerabilities
of knowing an application's internals.

- Firstly it checks whether the request object is available. If it is
not, this is an indicator that the request is in the early stages (e.g.
in the before request hook). In this case, the exception raised will not
be fatal as Dancer will not handle it cleanly.

- Secondly, rather than returning to Dancer and doing nothing under such
circumstances, it now returns (as plain text) the default unexpected
exception error text. If this can't be rendered (because it is too early
in the request) then Dancer handles the exception cleanly with its
default error handling.

In order to enable the testing of show_errors being enabled and disabled
within one test file, this commit also evaluates that setting per
message request rather than initialisation (at which point it is not
possible to change it without reloading the module).